+++ galay-mongo/base/mongo_log.h
@@ -1,184 +1,72 @@
+/**
+ * @file mongo_log.h
+ * @brief galay-mongo 独立日志入口与埋点宏
+ */
+
 #ifndef GALAY_MONGO_LOG_H
 #define GALAY_MONGO_LOG_H
 
-#include <memory>
-#include <mutex>
-#include <string>
+#include "galay-kernel/common/log_macro.h"
 
-#include <spdlog/sinks/basic_file_sink.h>
-#include <spdlog/sinks/stdout_color_sinks.h>
-#include <spdlog/spdlog.h>
-
-namespace galay::mongo
+namespace galay::mongo::detail
 {
+struct MongoLogTag;
+} // namespace galay::mongo::detail
 
-using MongoLoggerPtr = std::shared_ptr<spdlog::logger>;
-
-class MongoLog
+namespace galay::mongo::log
 {
-public:
-    static MongoLog* getInstance()
-    {
-        static MongoLog instance;
-        return &instance;
-    }
+/**
+ * @brief 设置 galay-mongo 的库级 logger
+ *
+ * @details 只影响 `MONGO_LOG_*` 宏产生的日志，不会启用 kernel、ssl、http
+ * 或其他 galay 库日志。推荐在创建 Mongo client 之前的单线程初始化阶段调用。
+ *
+ * @param logger 用户自定义 logger；传入 nullptr 时禁用 galay-mongo 日志。
+ */
+void set(::galay::kernel::BaseLogger::uptr logger);
 
-    static void enable()
-    {
-        console();
-    }
+/**
+ * @brief 获取 galay-mongo 当前 logger
+ *
+ * @return 当前 logger 指针；未设置时返回 nullptr。
+ *
+ * @note 返回指针由 `set()` 注入的 unique_ptr 管理，调用方不得释放。
+ */
+[[nodiscard]] ::galay::kernel::BaseLogger* get() noexcept;
+} // namespace galay::mongo::log
 
-    static void console()
-    {
-        console("MongoLogger");
-    }
+/// @brief 判断指定级别的 galay-mongo 日志是否会实际写入
+#define MONGO_LOG_ENABLED(level)                                                 \
+    GALAY_LOG_ENABLED(::galay::mongo::log::get, level)
 
-    static void console(const std::string& logger_name)
-    {
-        auto instance = getInstance();
-        std::lock_guard<std::mutex> lock(instance->m_mutex);
-        try {
-            auto logger = spdlog::get(logger_name);
-            if (!logger) {
-                logger = spdlog::stdout_color_mt(logger_name);
-            }
-            applyDefault(logger);
-            instance->m_logger = std::move(logger);
-        } catch (const spdlog::spdlog_ex&) {
-            instance->m_logger = spdlog::get(logger_name);
-        }
-    }
+/// @brief galay-mongo 追踪日志宏
+#define MONGO_LOG_TRACE(tag, ...)                                                \
+    GALAY_LOG_WITH_LOGGER(::galay::mongo::log::get,                              \
+                          ::galay::kernel::LogLevel::kTrace, "[mongo] " tag,     \
+                          __VA_ARGS__)
 
-    static void file(const std::string& log_file_path = "galay-mongo.log",
-                     const std::string& logger_name = "MongoLogger",
-                     bool truncate = false)
-    {
-        auto instance = getInstance();
-        std::lock_guard<std::mutex> lock(instance->m_mutex);
-        auto logger = std::make_shared<spdlog::logger>(
-            logger_name,
-            std::make_shared<spdlog::sinks::basic_file_sink_mt>(log_file_path, !truncate));
-        applyDefault(logger);
-        instance->m_logger = std::move(logger);
-    }
+/// @brief galay-mongo 调试日志宏
+#define MONGO_LOG_DEBUG(tag, ...)                                                \
+    GALAY_LOG_WITH_LOGGER(::galay::mongo::log::get,                              \
+                          ::galay::kernel::LogLevel::kDebug, "[mongo] " tag,     \
+                          __VA_ARGS__)
 
-    static void disable()
-    {
-        auto instance = getInstance();
-        std::lock_guard<std::mutex> lock(instance->m_mutex);
-        if (instance->m_logger) {
-            instance->m_logger->set_level(spdlog::level::off);
-        }
-        instance->m_logger.reset();
-    }
+/// @brief galay-mongo 信息日志宏
+#define MONGO_LOG_INFO(tag, ...)                                                 \
+    GALAY_LOG_WITH_LOGGER(::galay::mongo::log::get,                              \
+                          ::galay::kernel::LogLevel::kInfo, "[mongo] " tag,      \
+                          __VA_ARGS__)
 
-    static void setLogger(MongoLoggerPtr logger)
-    {
-        auto instance = getInstance();
-        std::lock_guard<std::mutex> lock(instance->m_mutex);
-        instance->m_logger = std::move(logger);
-    }
+/// @brief galay-mongo 警告日志宏
+#define MONGO_LOG_WARN(tag, ...)                                                 \
+    GALAY_LOG_WITH_LOGGER(::galay::mongo::log::get,                              \
+                          ::galay::kernel::LogLevel::kWarn, "[mongo] " tag,      \
+                          __VA_ARGS__)
 
-    MongoLoggerPtr getLogger() const
-    {
-        std::lock_guard<std::mutex> lock(m_mutex);
-        return m_logger;
-    }
-
-private:
-    static void applyDefault(const MongoLoggerPtr& logger)
-    {
-        if (!logger) {
-            return;
-        }
-        logger->set_pattern("[%Y-%m-%d %T.%e] [%^%L%$] [%s:%#] %v");
-#ifdef ENABLE_DEBUG
-        logger->set_level(spdlog::level::debug);
-#else
-        logger->set_level(spdlog::level::info);
-#endif
-    }
-
-private:
-    mutable std::mutex m_mutex;
-    MongoLoggerPtr m_logger;
-};
-
-class MongoLogger
-{
-public:
-    MongoLogger() = default;
-    explicit MongoLogger(MongoLoggerPtr logger)
-        : m_logger(std::move(logger))
-    {
-    }
-
-    void ensure(const std::string& logger_name)
-    {
-        if (m_logger) {
-            return;
-        }
-
-        m_logger = MongoLog::getInstance()->getLogger();
-        if (m_logger) {
-            return;
-        }
-
-        try {
-            m_logger = spdlog::get(logger_name);
-        } catch (const spdlog::spdlog_ex&) {
-            m_logger = nullptr;
-        }
-    }
-
-    void set(MongoLoggerPtr logger)
-    {
-        m_logger = std::move(logger);
-    }
-
-    MongoLoggerPtr get() const
-    {
-        return m_logger;
-    }
-
-    bool valid() const
-    {
-        return static_cast<bool>(m_logger);
-    }
-
-private:
-    MongoLoggerPtr m_logger;
-};
-
-namespace detail
-{
-inline MongoLoggerPtr resolveLogger(const MongoLoggerPtr& logger)
-{
-    if (logger) {
-        return logger;
-    }
-    return MongoLog::getInstance()->getLogger();
-}
-} // namespace detail
-
-} // namespace galay::mongo
-
-/// @name 日志宏（logger 为空时自动尝试全局 logger）.
-/// @{
-#define MongoLogTrace(logger, ...) \
-    do { auto _logger = ::galay::mongo::detail::resolveLogger((logger)); if (_logger) SPDLOG_LOGGER_TRACE(_logger, __VA_ARGS__); } while (0)
-
-#define MongoLogDebug(logger, ...) \
-    do { auto _logger = ::galay::mongo::detail::resolveLogger((logger)); if (_logger) SPDLOG_LOGGER_DEBUG(_logger, __VA_ARGS__); } while (0)
-
-#define MongoLogInfo(logger, ...) \
-    do { auto _logger = ::galay::mongo::detail::resolveLogger((logger)); if (_logger) SPDLOG_LOGGER_INFO(_logger, __VA_ARGS__); } while (0)
-
-#define MongoLogWarn(logger, ...) \
-    do { auto _logger = ::galay::mongo::detail::resolveLogger((logger)); if (_logger) SPDLOG_LOGGER_WARN(_logger, __VA_ARGS__); } while (0)
-
-#define MongoLogError(logger, ...) \
-    do { auto _logger = ::galay::mongo::detail::resolveLogger((logger)); if (_logger) SPDLOG_LOGGER_ERROR(_logger, __VA_ARGS__); } while (0)
-/// @}
+/// @brief galay-mongo 错误日志宏
+#define MONGO_LOG_ERROR(tag, ...)                                                \
+    GALAY_LOG_WITH_LOGGER(::galay::mongo::log::get,                              \
+                          ::galay::kernel::LogLevel::kError, "[mongo] " tag,     \
+                          __VA_ARGS__)
 
 #endif // GALAY_MONGO_LOG_H